flaskBack/requirements.txt
certifi==2020.12.5
click==7.1.2
Cython==0.29.21
DateTime==4.3
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.4
PyJWT==1.7.1
PyMySQL==0.10.1
pytz==2020.4
PyYAML==5.3.1
SQLAlchemy==1.3.20
Werkzeug==1.0.1
zope.interface==5.2.0
